Atom class at working example stage
The atom class needs to keep track of its atomic mass, its symbol, its
name and its atomic number, in that order of importance. At some point,
this class (and others) should retrieve the data automatically from a
database, but for now, a hack-ish function called setProperties is
provided to fill in the missing numbers.