CHNO system available for testing
An instance of the atom class can be created using one of the C, H, N, O
elements for quick testing. The setProperties() function has been
removed.

Atom objects can now self-report. Calling the report() function prints
out a quick report of all the properties of the atom. This is in
addition to the ostream interface.