Updated cmake for the project.
The CMakeLists.txt files in the project were updated to give the user more
options about what/where/how to install things for the equilibrium project.
Users should do the following to build the project now (from the main project
directory):

   1. mkdir build
   2. ccmake ../trunk
   3. configure any options present
   4. make install

A couple of notes about the changes made in this commit:

   1. The C++ source code is developed as a library, from which executables
      (unit tests, future main programs, etc.) will be linked. This is more
      flexible because folks might just want to link this equilibrium library to
      more complex/different codes in the future...not just the GUI that will be
      developed here.
   2. All source code, CMakeLists.txt, etc. are pushed to trunk to keep things
      cleaner. Users can now make a build directory and many different
      installation directories (perhaps for different compilers) in the main
      project directory without getting the source contaminated.
   3. Unit tests for classes, etc. should be put in the tests directory. Going
      with this strategy will allow us to make many unit tests instead of just
      one main.cpp.

TODO: for some reason I couldn't get INCLUDE_DIRECTORIES() working properly in
cmake tonight. If you take a look at trunk/tests/atom_molecule.cpp you will
notice that the includes at the top of the file are pretty nasty (relative paths
to the actual header files). This should not have to be done, and can be
remedied with some proper cmake settings...we just need to figure them out!